+++ cmd/root.go
@@ -48,6 +48,8 @@ var (
 	githubAppInstallationID      string
 	githubUserOption             string
 	githubFilterLanguage         string
+	cronTimerMinutes             string
+	recloneServerPort            string
 	includeSubmodules            bool
 	skipArchived                 bool
 	skipForks                    bool
@@ -73,7 +75,6 @@ var (
 	ghorgPreserveScmHostname     bool
 	ghorgPruneUntouched          bool
 	ghorgPruneUntouchedNoConfirm bool
-	args                         []string
 	cloneErrors                  []string
 	cloneInfos                   []string
 )
@@ -168,6 +169,10 @@ func getOrSetDefaults(envVar string) {
 			os.Setenv(envVar, "false")
 		case "GHORG_NO_CLEAN":
 			os.Setenv(envVar, "false")
+		case "GHORG_CRON_TIMER_MINUTES":
+			os.Setenv(envVar, "60")
+		case "GHORG_RECLONE_SERVER_PORT":
+			os.Setenv(envVar, ":8080")
 		case "GHORG_FETCH_ALL":
 			os.Setenv(envVar, "false")
 		case "GHORG_DRY_RUN":
@@ -300,6 +305,8 @@ func InitConfig() {
 	getOrSetDefaults("GHORG_EXIT_CODE_ON_CLONE_INFOS")
 	getOrSetDefaults("GHORG_EXIT_CODE_ON_CLONE_ISSUES")
 	getOrSetDefaults("GHORG_STATS_ENABLED")
+	getOrSetDefaults("GHORG_CRON_TIMER_MINUTES")
+	getOrSetDefaults("GHORG_RECLONE_SERVER_PORT")
 	// Optionally set
 	getOrSetDefaults("GHORG_TARGET_REPOS_PATH")
 	getOrSetDefaults("GHORG_CLONE_DEPTH")
@@ -407,7 +414,11 @@ func init() {
 	lsCmd.Flags().BoolP("long", "l", false, "Display detailed information about each clone directory, including size and number of repositories. Note: This may take longer depending on the number and size of the cloned organizations.")
 	lsCmd.Flags().BoolP("total", "t", false, "Display total amounts of all repos cloned. Note: This may take longer depending on the number and size of the cloned organizations.")
 
-	rootCmd.AddCommand(lsCmd, versionCmd, cloneCmd, reCloneCmd, examplesCmd)
+	recloneCronCmd.Flags().StringVarP(&cronTimerMinutes, "minutes", "m", "", "GHORG_CRON_TIMER_MINUTES - Number of minutes to run the reclone command on a cron")
+
+	recloneServerCmd.Flags().StringVarP(&recloneServerPort, "port", "p", "", "GHORG_RECLONE_SERVER_PORT - Specifiy the port the reclone server will run on.")
+
+	rootCmd.AddCommand(lsCmd, versionCmd, cloneCmd, reCloneCmd, examplesCmd, recloneServerCmd, recloneCronCmd)
 }
 
 func Execute() {